refactor(closing): extract judge tests into a separate file
Move the large inline test module from judge.rs into its own file to reduce the module's length and improve readability. The tests are unchanged in behaviour.

Auto-committed-on: dragonfly
